pyproject.toml
[project]
name = "athena_dr"
dynamic = ["version"]
description = "Deep Research assistant"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiohttp>=3.13.2",
    "beautifulsoup4>=4.14.3",
    "blobfile>=3.1.0",
    "cohere>=5.20.0",
    "crawl4ai>=0.7.8",
    "datasets>=4.4.1",
    "diskcache>=5.6.3",
    "fastmcp>=2.13.3",
    "litellm>=1.80.9",
    "lxml>=5.4.0",
    "nltk>=3.9.2",
    "omegaconf>=2.3.0",
    "openai>=2.9.0",
    "pdfplumber>=0.11.7",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "retry>=0.9.2",
    "tenacity>=9.1.2",
    "tinydb>=4.8.2",
    "tinydb-serialization>=2.2.0",
    "transformers>=4.57.3",
    "typer>=0.20.0",
    "weave>=0.52.22",
]

[dependency-groups]
dev = [
    "uv>=0.4.20",
    "pytest>=8.3.3",
    "isort>=5.13.2",
    "black>=24.10.0",
    "ruff>=0.6.9",
]
docs = [
    "mkdocs>=1.6.1",
    "mkdocstrings>=0.26.1",
    "mkdocstrings-python>=1.11.1",
    "mkdocs-material>=9.5.39",
    "mkdocs-minify-plugin>=0.8.0",
    "mkdocs-glightbox>=0.4.0",
    "mkdocs-jupyter>=0.25.0",
    "mkdocs-panzoom-plugin",
    "jupyter>=1.1.1",
]

[project.scripts]
athena-dr-mcp = "athena_dr.agent.mcp_backend.main:main"
athena-dr-list-tools = "athena_dr.utils:list_tools_from_server"
athena-dr-workflows = "athena_dr.agent.workflows.cli:launch_web_page_reader_cli"
athena-dr = "athena_dr.agent.workflows.cli:launch_auto_search_workflow_cli"
athena-dr-evals = "athena_dr.evals.cli:main"

[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[tool.hatch.version]
source = "vcs"

[tool.hatch.build.targets.wheel]
packages = ["athena_dr"]

[tool.ruff]
target-version = "py312"

[tool.ruff.lint]
select = ["I", "F401"]

[tool.ruff.lint.isort]
known-first-party = ["athena_dr"]
section-order = ["future", "standard-library", "third-party", "first-party", "local-folder"]
combine-as-imports = true

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
